fix(ci): satisfy runtime image lint gates
